refactor: extract table Alignment and Pinning into a common schema
Alignment is a table primitive, not a unified-report concept: @ui/Table,
the DataTable blocks, the general ledger tables, and the bank transactions
table all read it, so blocks and ui were importing a report domain schema.
Both enums and their transformed schemas move to schemas/common/table, and
unifiedReport imports them from there.
