+++ src/components/features/generalLedger/JournalTable/JournalTable.tsx
@@ -2,9 +2,9 @@ import { useCallback, useContext, useLayoutEffect, useMemo } from 'react'
 import { type Row } from '@tanstack/react-table'
 import { useTranslation } from 'react-i18next'
 
+import { Alignment } from '@schemas/common/table'
 import { LedgerEntryDirection } from '@schemas/generalLedger/ledgerAccount'
 import { type LedgerEntry } from '@schemas/generalLedger/ledgerEntry'
-import { Alignment } from '@schemas/reports/unifiedReport'
 import { humanizeEnum } from '@utils/format'
 import { entryNumber, sumLineItemAmountsByDirection } from '@utils/journal'
 import { JOURNAL_PAGE_SIZE } from '@hooks/legacy/useJournal'
