+++ src/components/features/generalLedger/LedgerAccountLineItemsTable/LedgerAccountLineItemsTable.tsx
@@ -4,10 +4,10 @@ import type { TFunction } from 'i18next'
 import { useTranslation } from 'react-i18next'
 
 import { LedgerAccountNodeType } from '@internal-types/chartOfAccounts'
+import { Alignment } from '@schemas/common/table'
 import { LedgerEntryDirection } from '@schemas/generalLedger/ledgerAccount'
 import { type LedgerAccountLineItem } from '@schemas/generalLedger/ledgerEntry'
 import { decodeLedgerEntrySource } from '@schemas/generalLedger/ledgerEntrySource'
-import { Alignment } from '@schemas/reports/unifiedReport'
 import { lineEntryNumber } from '@utils/journal'
 import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
 import { LedgerAccountsContext } from '@contexts/LedgerAccountsContext/LedgerAccountsContext'
