+++ src/components/features/generalLedger/LedgerEntryDetails/LedgerEntryDetailsLineItemsTable.tsx
@@ -2,9 +2,9 @@ import { useMemo } from 'react'
 import { type Row } from '@tanstack/react-table'
 import { useTranslation } from 'react-i18next'
 
+import { Alignment } from '@schemas/common/table'
 import { LedgerEntryDirection } from '@schemas/generalLedger/ledgerAccount'
 import { type LedgerEntry } from '@schemas/generalLedger/ledgerEntry'
-import { Alignment } from '@schemas/reports/unifiedReport'
 import { sumLineItemAmountsByDirection } from '@utils/journal'
 import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
 import { Badge, BadgeVariant } from '@ui/Badge/Badge'
