+++ src/components/features/unifiedReports/utils.tsx
@@ -1,6 +1,7 @@
 import { type Row } from '@tanstack/react-table'
 
-import { Pinning, type UnifiedReportColumn, type UnifiedReportRow } from '@schemas/reports/unifiedReport'
+import { Pinning } from '@schemas/common/table'
+import { type UnifiedReportColumn, type UnifiedReportRow } from '@schemas/reports/unifiedReport'
 import { asMutable } from '@utils/asMutable'
 import { type ColumnNode, type GroupColumn, type LeafColumn } from '@blocks/DataTable/utils/column/nesting'
 import { type ColumnPinningSide } from '@blocks/DataTable/utils/column/pinning'
